feat(cli): trueno-rag and trueno-zram had no route through apr at all
Both crates declared their entire command surface inside a main.rs. A
command enum in a binary target is importable by nothing, so the
standalone binary was the only way to reach any of it -- ten commands
with no route through apr. Retiring those binaries, as APR-MONO intends,
would have DELETED the capability rather than relocating it.

Each command surface moves into its crate's lib as `pub enum Commands`
plus `pub fn run()` (parse) and `pub fn dispatch()` (execute); the binary
becomes a shim over `run()`. apr's new arm calls the SAME `dispatch`, so
`apr rag index` and `trueno-rag index` cannot drift -- verified by
running both against one index and diffing: byte-identical output.

Exposing zram surfaced a defect that had never been hit. `pages` derived
short `-p` while `pattern` explicitly claimed `-p`; clap catches that in
a debug_assert that fires before any argument is parsed, so EVERY
`trueno-zram benchmark` invocation -- `--help` included -- aborted with
exit 101. Nothing ever ran the command, so nothing noticed. `pages` now
takes `-n`.

Evidence, on this box:
  apr rag info      -> reads LoaderRegistry, lists txt, md, srt, vtt
  apr rag index/query -> indexes 2 docs, ranks the matching one first
  apr zram status   -> reports the machine's real zram0/zram1
  apr zram benchmark --pages 200 --algorithm lz4
                    -> Lz4 Avx512 0.14 GB/s, ratio 3.87x

Five tests in beat_apr_rag_zram_reach.rs, each mutation-verified:

  * revert `-n` to the derived `-p`
      -> only apr_zram_benchmark_runs_instead_of_panicking goes RED

  * short-circuit dispatch_sibling_cli_commands to None
      -> the two help-listing tests stay GREEN and all three execution
         tests go RED

That second split is the point. The enum stays fully wired and `--help`
is unchanged, so a reach test built only on help output would have
reported green while nothing executed.

The cross-binary byte-identity check is deliberately NOT a test:
CARGO_BIN_EXE_* exposes only this package's binaries, so writing it from
apr-cli yields an apr-vs-apr comparison -- an oracle that agrees with
itself by construction.

Refs #2494
