+++ crates/aprender-zram-cli/src/commands/remove.rs
@@ -6,7 +6,7 @@ use clap::Args;
 use trueno_zram_core::zram::{SysfsOps, ZramOps};
 
 /// Arguments for removing a zram device.
-#[derive(Args)]
+#[derive(Debug, Args)]
 pub struct RemoveArgs {
     /// Device number to remove.
     #[arg(short, long, default_value = "0")]
@@ -18,6 +18,9 @@ pub struct RemoveArgs {
 }
 
 /// Remove a zram device.
+///
+/// # Errors
+/// Returns an error if the zram device cannot be removed.
 pub fn remove(args: &RemoveArgs) -> Result<(), Box<dyn std::error::Error>> {
     let ops = SysfsOps::new();
     ops.remove(args.device, args.force)?;
