+++ crates/aprender-zram-cli/src/main.rs
@@ -1,58 +1,13 @@
-//! trueno-zram CLI - zramctl replacement with SIMD acceleration.
+//! trueno-zram binary. The command surface lives in the library
+//! (`aprender_zram_cli`) so `apr zram` can reach the same code; see the module
+//! docs there.
 
 #![deny(missing_docs)]
 #![deny(clippy::panic)]
 #![warn(clippy::all, clippy::pedantic)]
 
-mod commands;
-mod output;
-
-use clap::{Parser, Subcommand};
 use std::process::ExitCode;
 
-/// trueno-zram: SIMD-accelerated zram management
-#[derive(Parser)]
-#[command(name = "trueno-zram")]
-#[command(author, version, about, long_about = None)]
-struct Cli {
-    /// Output format
-    #[arg(long, default_value = "table")]
-    format: output::OutputFormat,
-
-    #[command(subcommand)]
-    command: Commands,
-}
-
-#[derive(Subcommand)]
-enum Commands {
-    /// Create and configure a zram device
-    Create(commands::CreateArgs),
-
-    /// Remove a zram device
-    Remove(commands::RemoveArgs),
-
-    /// Show zram device status
-    Status(commands::StatusArgs),
-
-    /// Run compression benchmarks
-    Benchmark(commands::BenchmarkArgs),
-}
-
 fn main() -> ExitCode {
-    let cli = Cli::parse();
-
-    let result = match &cli.command {
-        Commands::Create(args) => commands::create(args),
-        Commands::Remove(args) => commands::remove(args),
-        Commands::Status(args) => commands::status(args, cli.format),
-        Commands::Benchmark(args) => commands::benchmark(args),
-    };
-
-    match result {
-        Ok(()) => ExitCode::SUCCESS,
-        Err(e) => {
-            eprintln!("Error: {e}");
-            ExitCode::FAILURE
-        }
-    }
+    aprender_zram_cli::run()
 }
